just added .toDouble() and modified delay so bacteria live longer during a test

+++ partie4/src/Lab/SwarmBacterium.cpp
@@ -27,7 +27,7 @@ void SwarmBacterium::move(sf::Time dt){
     Vec2d movement = stepDiffEq(getPosition(),getSpeedVector(),dt,(*this),DiffEqAlgorithm::EC).position - getPosition();
     if(movement.lengthSquared()>0.001){
     (*this).CircularBody::move(movement);
-    consumeEnergy((movement).length()*getEnergyConsumption());
+    consumeEnergy((movement).length()*getEnergyConsumption().toDouble());
     }
     if(swarm->isLeader(this)){
         tumble();
